fix and reenable uniforme2e
