+++ iceberg/src/test/scala/org/apache/spark/sql/delta/ConvertToIcebergSuite.scala
@@ -110,6 +110,7 @@ class ConvertToIcebergSuite extends QueryTest with Eventually {
         s"""CREATE TABLE `${testTableName}` (col1 INT) USING DELTA
            |TBLPROPERTIES (
            |  'delta.columnMapping.mode' = 'name',
+           |  'delta.enableIcebergCompatV2' = 'true',
            |  'delta.universalFormat.enabledFormats' = 'iceberg'
            |)""".stripMargin)
       runDeltaSql(s"INSERT INTO `$testTableName` VALUES (123)")
@@ -156,6 +157,7 @@ class ConvertToIcebergSuite extends QueryTest with Eventually {
   def withDefaultTablePropsInSQLConf(f: => Unit): Unit = {
     withSQLConf(
       DeltaConfigs.COLUMN_MAPPING_MODE.defaultTablePropertyKey -> "name",
+      DeltaConfigs.ICEBERG_COMPAT_V1_ENABLED.defaultTablePropertyKey -> "true",
       DeltaConfigs.UNIVERSAL_FORMAT_ENABLED_FORMATS.defaultTablePropertyKey -> "iceberg"
     ) { f }
   }
