+++ iceberg/src/test/scala/org/apache/spark/sql/delta/uniform/UniFormE2EIcebergSuite.scala
@@ -1,0 +1,70 @@
+/*
+ * Copyright (2021) The Delta Lake Project Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.delta.uniform
+
+import org.apache.spark.sql.delta.sources.DeltaSQLConf
+import org.apache.spark.{SparkConf, SparkSessionSwitch}
+import org.apache.spark.sql.{Row, SparkSession}
+import org.apache.spark.sql.delta.test.DeltaSQLCommandTest
+import org.apache.spark.sql.delta.uniform.hms.HMSTest
+
+/**
+ * This trait allows the tests to write with Delta
+ * using a in-memory HiveMetaStore as catalog,
+ * and read from the same HiveMetaStore with Iceberg.
+ */
+trait WriteDeltaHMSReadIceberg extends UniFormE2ETest
+  with DeltaSQLCommandTest with HMSTest with SparkSessionSwitch {
+
+  override protected def sparkConf: SparkConf =
+    setupSparkConfWithHMS(super.sparkConf)
+      .set(DeltaSQLConf.DELTA_UNIFORM_ICEBERG_SYNC_CONVERT_ENABLED.key, "true")
+      .set(DeltaSQLConf.DELTA_OPTIMIZE_WRITE_ENABLED.key, "true")
+      .set(DeltaSQLConf.DELTA_AUTO_COMPACT_ENABLED.key, "true")
+
+  private var _readerSparkSession: Option[SparkSession] = None
+  /**
+   * Verify the result by reading from the reader session and compare the result to the expected.
+   *
+   * @param table  write table name
+   * @param fields fields to verify, separated by comma. E.g., "col1, col2"
+   * @param orderBy fields to order the results, separated by comma.
+   * @param expect expected result
+   */
+  protected override def readAndVerify(
+      table: String, fields: String, orderBy: String, expect: Seq[Row]): Unit = {
+    val translated = tableNameForRead(table)
+    withSession(readerSparkSession) { session =>
+      checkAnswer(session.sql(s"SELECT $fields FROM $translated ORDER BY $orderBy"), expect)
+    }
+  }
+
+  protected def readerSparkSession: SparkSession = {
+    if (_readerSparkSession.isEmpty) {
+      // call to newSession makes sure
+      // [[SparkSession.getOrCreate]] gives a new session
+      // and [[SparkContext.getOrCreate]] uses a new context
+      _readerSparkSession = Some(newSession(createIcebergSparkSession))
+    }
+    _readerSparkSession.get
+  }
+}
+
+/**
+ * No test should go here. Please add tests in [[UniFormE2EIcebergSuiteBase]]
+ */
+class UniFormE2EIcebergSuite extends UniFormE2EIcebergSuiteBase with WriteDeltaHMSReadIceberg
