+++ spark/src/test/scala/org/apache/spark/sql/delta/uniform/UniFormE2EIcebergSuiteBase.scala
@@ -16,14 +16,14 @@
 
 package org.apache.spark.sql.delta.uniform
 
-import org.apache.spark.sql.delta.DeltaLog
-import org.apache.spark.sql.delta.DeltaTestUtils.withTimeZone
-import org.apache.spark.sql.delta.sources.DeltaSQLConf.UTC_TIMESTAMP_PARTITION_VALUES
-
 import org.apache.spark.sql.Row
-import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.types._
 
+/**
+ * This test suite base aims at testing the end-to-end behavior of UniForm.
+ * It writes Delta tables, and reads the generated Iceberg tables to
+ * perform verification.
+ */
 abstract class UniFormE2EIcebergSuiteBase extends UniFormE2ETest {
 
   val testTableName = "delta_table"
@@ -42,6 +42,91 @@ abstract class UniFormE2EIcebergSuiteBase extends UniFormE2ETest {
     }
   }
 
+  test("CIUD") {
+    withTable(testTableName) {
+      write(
+        s"""CREATE TABLE `$testTableName` (col1 INT) USING DELTA
+           |TBLPROPERTIES (
+           |  'delta.columnMapping.mode' = 'name',
+           |  'delta.enableIcebergCompatV1' = 'true',
+           |  'delta.universalFormat.enabledFormats' = 'iceberg'
+           |)""".stripMargin)
+      write(s"INSERT INTO `$testTableName` VALUES (123),(456),(567),(331)")
+      write(s"UPDATE `$testTableName` SET col1 = 191 WHERE col1 = 567")
+      write(s"DELETE FROM `$testTableName` WHERE col1 = 456")
+
+      readAndVerify(testTableName, "col1", "col1", Seq(Row(123), Row(191), Row(331)))
+    }
+  }
+
+  test("Nested struct schema test") {
+    withTable(testTableName) {
+      write(s"""CREATE TABLE $testTableName
+           | (col1 INT, col2 STRUCT<f1: STRUCT<f2: INT, f3: STRUCT<f4: INT, f5: INT>
+           | , f6: INT>, f7: INT>) USING DELTA
+           |TBLPROPERTIES (
+           |  'delta.columnMapping.mode' = 'name',
+           |  'delta.enableIcebergCompatV1' = 'true',
+           |  'delta.universalFormat.enabledFormats' = 'iceberg'
+           |)""".stripMargin)
+
+      val data = Seq(
+        Row(1, Row(Row(2, Row(3, 4), 5), 6))
+      )
+
+      val innerStruct3 = StructType(
+          StructField("f4", IntegerType) ::
+            StructField("f5", IntegerType) :: Nil)
+
+      val innerStruct2 = StructType(
+        StructField("f2", IntegerType) ::
+          StructField("f3", innerStruct3) ::
+          StructField("f6", IntegerType) :: Nil)
+
+      val innerStruct = StructType(
+        StructField("f1", innerStruct2) ::
+          StructField("f7", IntegerType) :: Nil)
+
+      val schema = StructType(
+        StructField("col1", IntegerType) ::
+          StructField("col2", innerStruct) :: Nil)
+
+      val tableFullName = tableNameForRead(testTableName)
+
+      spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
+        .write.format("delta").mode("append")
+        .saveAsTable(testTableName)
+
+      readAndVerify(tableFullName, "col1, col2", "col1", data)
+    }
+  }
+
+  test("Re-enable test") {
+    withTable(testTableName) {
+      write(
+        s"""CREATE TABLE $testTableName (col1 INT) USING DELTA
+           |TBLPROPERTIES (
+           |  'delta.columnMapping.mode' = 'name',
+           |  'delta.enableIcebergCompatV1' = 'true',
+           |  'delta.universalFormat.enabledFormats' = 'iceberg'
+           |)""".stripMargin)
+      write(s"INSERT INTO $testTableName VALUES (1)")
+      readAndVerify(testTableName, "col1", "col1", Seq(Row(1)))
+
+      write(s"ALTER TABLE `$testTableName` UNSET TBLPROPERTIES " +
+        s"('delta.universalFormat.enabledFormats')")
+      write(s"""
+               | REORG TABLE $testTableName APPLY
+               | (UPGRADE UNIFORM (ICEBERG_COMPAT_VERSION = 2))
+               |""".stripMargin)
+      write(s"INSERT INTO $testTableName VALUES (2)")
+      readAndVerify(testTableName, "col1", "col1", Seq(Row(1), Row(2)))
+    }
+  }
+
+  // TODO createReaderSparkSession is no longer supported.
+  // Please use readAndVerify and re-enable the cases
+  /*
   test("Insert Partitioned Table") {
     val partitionColumns = Array(
       "str STRING",
@@ -144,88 +229,5 @@ abstract class UniFormE2EIcebergSuiteBase extends UniFormE2ETest {
       }
     }
   }
-
-  test("CIUD") {
-    withTable(testTableName) {
-      write(
-        s"""CREATE TABLE `$testTableName` (col1 INT) USING DELTA
-           |TBLPROPERTIES (
-           |  'delta.columnMapping.mode' = 'name',
-           |  'delta.enableIcebergCompatV1' = 'true',
-           |  'delta.universalFormat.enabledFormats' = 'iceberg'
-           |)""".stripMargin)
-      write(s"INSERT INTO `$testTableName` VALUES (123),(456),(567),(331)")
-      write(s"UPDATE `$testTableName` SET col1 = 191 WHERE col1 = 567")
-      write(s"DELETE FROM `$testTableName` WHERE col1 = 456")
-
-      readAndVerify(testTableName, "col1", "col1", Seq(Row(123), Row(191), Row(331)))
-    }
-  }
-
-  test("Nested struct schema test") {
-    withTable(testTableName) {
-      write(s"""CREATE TABLE $testTableName
-           | (col1 INT, col2 STRUCT<f1: STRUCT<f2: INT, f3: STRUCT<f4: INT, f5: INT>
-           | , f6: INT>, f7: INT>) USING DELTA
-           |TBLPROPERTIES (
-           |  'delta.columnMapping.mode' = 'name',
-           |  'delta.enableIcebergCompatV1' = 'true',
-           |  'delta.universalFormat.enabledFormats' = 'iceberg'
-           |)""".stripMargin)
-
-      val data = Seq(
-        Row(1, Row(Row(2, Row(3, 4), 5), 6))
-      )
-
-      val innerStruct3 = StructType(
-          StructField("f4", IntegerType) ::
-            StructField("f5", IntegerType) :: Nil)
-
-      val innerStruct2 = StructType(
-        StructField("f2", IntegerType) ::
-          StructField("f3", innerStruct3) ::
-          StructField("f6", IntegerType) :: Nil)
-
-      val innerStruct = StructType(
-        StructField("f1", innerStruct2) ::
-          StructField("f7", IntegerType) :: Nil)
-
-      val schema = StructType(
-        StructField("col1", IntegerType) ::
-          StructField("col2", innerStruct) :: Nil)
-
-      val tableFullName = tableNameForRead(testTableName)
-
-      spark.createDataFrame(spark.sparkContext.parallelize(data), schema)
-        .write.format("delta").mode("append")
-        .saveAsTable(testTableName)
-
-      val result = read(s"SELECT * FROM $tableFullName")
-
-      assert(result.head === data.head)
-    }
-  }
-
-  test("Re-enable test") {
-    withTable(testTableName) {
-      write(
-        s"""CREATE TABLE $testTableName (col1 INT) USING DELTA
-           |TBLPROPERTIES (
-           |  'delta.columnMapping.mode' = 'name',
-           |  'delta.enableIcebergCompatV1' = 'true',
-           |  'delta.universalFormat.enabledFormats' = 'iceberg'
-           |)""".stripMargin)
-      write(s"INSERT INTO $testTableName VALUES (1)")
-      readAndVerify(testTableName, "col1", "col1", Seq(Row(1)))
-
-      write(s"ALTER TABLE `$testTableName` UNSET TBLPROPERTIES " +
-        s"('delta.universalFormat.enabledFormats')")
-      write(s"""
-               | REORG TABLE $testTableName APPLY
-               | (UPGRADE UNIFORM (ICEBERG_COMPAT_VERSION = 2))
-               |""".stripMargin)
-      write(s"INSERT INTO $testTableName VALUES (2)")
-      readAndVerify(testTableName, "col1", "col1", Seq(Row(1), Row(2)))
-    }
-  }
+*/
 }
